Revert "Push Pop and Code After B Instruction"
