--- SeniorProject2025/PushPopErrors.h
@@ -1,16 +1,0 @@
-#pragma once
-
-// PushPopErrors.h
-
-#ifndef PUSHPOPERRORS_h
-#define PUSHPOPERRORS_h
-
-#include <iostream>
-#include <fstream>
-#include <string>
-#include <stack>
-using namespace std;
-
-int detectPushPopMismatch(const string&);
-
-#endif
