+++ SeniorProject2025/SeniorProject2025.vcxproj.filters
@@ -30,9 +30,6 @@
     <ClCompile Include="calculations.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="PushPopErrors.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="main.h">
@@ -51,9 +48,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="flags.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="PushPopErrors.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
