+++ SeniorProject2025/branchAndSubroutines.cpp
@@ -1,159 +1,13 @@
 // branchAndSubroutines.cpp
 
 #include "branchAndSubroutines.h"
-#include "flags.h"
 
 // SUBROUTINES
 //------------------------------------------------------------>
 
-// Branching and Line Number
-vector<Subroutine> subroutines;
-vector<SubroutineCall> subroutineCalls;
-unordered_set<string> userFunctions;
-unordered_set<string> systemCalls = { "printf", "scanf" };
-unordered_map<string, int> labelToLine;
-string label, currentSubroutine;
-int subroutineStart = 0;
-bool insideFunction = false;
-
-// Function to read file and gather subroutines and subroutine calls
-int processSubroutine(const string& filename) {
-
-	ifstream file(filename); // open file
-
-	if (!file.is_open()) {
-		cerr << "Error. File not opened: " << filename << endl;
-		return 0;
-	}
-
-	else {
-
-		bool insideBlockComment = false; // used to ignore block comments
-
-		string line;
-		int lineCount = 0;
-		bool branchDetected = false;
-		bool lrSaved = false;
-		while (getline(file, line)) {
-			if (!line.empty() && line.back() == '\r')
-				line.pop_back();
-
-			lineCount++;
-
-			// Ignore Comments and Empty Lines
-			if (!isCommentOrEmpty(line, insideBlockComment)) {
-
-				// SUBROUTINE CALLS BY LINE AND SUBROUTINE ERRORS
-				//------------------------------------------------------------>
-				// If subroutine exists on the line
-				string subroutineName;
-				if (findSubroutine(line, subroutineName)) {
-
-					// If a function was being tracked and lacks a return, report it
-					if (insideFunction && subroutines.back().makesBLCall && !subroutines.back().hasReturn)
-						cerr << "[ERROR] Function " << currentSubroutine
-						<< " (starting at line " << subroutineStart
-						<< ") does not return properly (missing BX LR or MOV PC, LR)\n";
-
-					currentSubroutine = subroutineName;
-					userFunctions.insert(currentSubroutine);
-					subroutines.push_back({ currentSubroutine, subroutineStart, lineCount - 1, false, false }); // add subroutine
-					subroutineStart = lineCount;
-					insideFunction = true;
-				}
-
-				// Detect if a BL call is made
-				string calledFunction;
-				if (insideFunction && findBLCall(line, calledFunction))
-
-					// Ignore excluded functions (e.g., printf, scanf)
-					if (excludedFunctions.find(calledFunction) == excludedFunctions.end())
-						subroutines.back().makesBLCall = true;
-
-				// Check if the function has a return instruction
-				if (insideFunction && isReturnInstruction(line))
-					subroutines.back().hasReturn = true;
-
-				// Store label positions
-				if (findSubroutine(line, label)) labelToLine[label] = lineCount;
-
-				// Check if LR is being saved
-				if (isSavingLR(line)) lrSaved = true;
-
-				// Check if LR is being restored
-				else if (isRestoringLR(line)) lrSaved = false;
-
-				// Detect BL instruction
-				else if (isBLInstruction(line)) {
-					// Extract the subroutine name being called
-					smatch match;
-					regex_search(line, match, std::regex(R"(\bBL\s+(\w+))"));
-					string calledFunction = match[1];
-
-					// Exclude system calls (like printf and scanf)
-					if (systemCalls.count(calledFunction) == 0) {
-						if (!lrSaved) {
-							cout << "Error: BL call to " << calledFunction
-								<< " in subroutine " << currentSubroutine
-								<< " without saving LR at line " << lineCount << ": "
-								<< line << endl;
-						}
-						lrSaved = false;  // Reset LR save status after a call
-					}
-				}
-
-				//------------------------------------------------------------<
-
-				// DETECT CODE AFTER BRANCH
-				//------------------------------------------------------------>
-				if (isBranchInstruction(line) &&
-					!line.find("printf") && !line.find("scanf")) {
-					branchDetected = true;
-				}
-
-				// If branch was detected and next line is executable code without a label
-				else if (branchDetected && isExecutableCode(line)) {
-					cout << "Error: Executable code after branch (no label) at line " << lineCount << ": " << line << std::endl;
-					branchDetected = false;
-				}
-
-				// Reset branch detection if a label is encountered
-				else if (isLabel(line)) branchDetected = false;
-				//------------------------------------------------------------<
-			}
-		}
-
-		// Check the last function in case it doesn't return properly
-		if (insideFunction && subroutines.back().makesBLCall && !subroutines.back().hasReturn)
-			cout << "[ERROR] Function " << currentSubroutine
-			<< " (starting at line " << subroutineStart
-			<< ") does not return properly (missing BX LR or MOV PC, LR)\n";
-
-		file.clear();
-		file.seekg(0); // restart file from beginning
-		lineCount = 0; // reset line count
-
-		// second file read
-		while (getline(file, line)) {
-			lineCount++;
-
-			// Find subroutine calls with a second pass
-			string instruction, target;
-			if (findSubroutineCall(line, instruction, target))
-				subroutineCalls.push_back({ lineCount, instruction, target });
-
-			//END OF SECOND FILE READ
-		}
-
-		printSubroutineCalls();
-
-		return 1;
-	}
-
-}
-
 // Function to print subroutines
-void printSubroutineCalls() {
+void printSubroutineCalls(vector<Subroutine> subroutines,
+	vector<SubroutineCall> subroutineCalls, unordered_map<string, int> labelToLine) {
 
 	// print subroutine calls and associated errors
 	cout << "\n >--- Subroutine Calls ---<\n";
@@ -242,6 +96,20 @@ bool findBLCall(const string& line, string& functionName) {
 	}
 
 	return false;
+}
+
+// Function to detect whether LR is saved
+bool isLRSaved(const string& line) {
+	regex saveLRRegexLow(R"(\b(push\s*\{\s*lr\s*\}|\bstmfd\s+sp!,\s*\{lr\})\b)", regex::icase);
+	regex saveLRRegexUp(R"(\b(PUSH\s*\{\s*LR\s*\}|\bSTMFD\s+SP!,\s*\{LR\})\b)", regex::icase);
+	return regex_search(line, saveLRRegexLow) || regex_search(line, saveLRRegexUp);
+}
+
+// Function to detect whether a return statement (BX LR or MOV PC, LR) is present
+bool isReturnInstruction(const string& line) {
+	regex returnRegexLow(R"(\b(bx\s+lr|mov\s+pc,\s*lr)\b)", regex::icase);
+	regex returnRegexUp(R"(\b(BX\s+LR|MOV\s+PC,\s*LR)\b)", regex::icase);
+	return regex_search(line, returnRegexLow) || regex_search(line, returnRegexUp);
 }
 
 //------------------------------------------------------------<
